requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==1.4.47
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
jinja2==3.1.2
aiofiles==23.1.0
pydantic==1.10.7
google-auth==2.19.1
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
nltk==3.8.1
email-validator==2.0.0
